requirements.txt
anyio==3.3.4
asgiref==3.4.1
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
fastapi==0.70.0
h11==0.12.0
httptools==0.2.0
idna==3.3
pycodestyle==2.8.0
pydantic==1.8.2
python-dotenv==0.19.1
PyYAML==6.0
requests==2.26.0
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
watchgod==0.7
boto3
moviepy
vosk
sklearn
face_recognition